[WIP] Created calendar event functionality and separatedform functionality into different files
