+++ src/components/calendar/Home.jsx
@@ -3,6 +3,18 @@ import Calendar from "@toast-ui/react-calendar";
 import "@toast-ui/calendar/dist/toastui-calendar.min.css";
 import "./HomeStyles.css";
 
+// Import utilities and components
+import { authAPI, calendarAPI, eventsAPI } from "./utils/api";
+import {
+  determineCalendarId,
+  transformCalendarData,
+  getEventStats,
+  formatCurrentDate,
+  getCalendarOptions,
+} from "./utils/calendarUtils";
+import CreateEventModal from "./CreateEventModal";
+import EventDetailModal from "./EventDetailModal";
+
 const Home = () => {
   // State management
   const [user, setUser] = useState(null);
@@ -33,14 +45,8 @@ const Home = () => {
   // Fetch current user
   const fetchUser = async () => {
     try {
-      const response = await fetch("http://localhost:8080/auth/me", {
-        credentials: "include",
-      });
-
-      if (response.ok) {
-        const userData = await response.json();
-        setUser(userData.user);
-      }
+      const userData = await authAPI.getMe();
+      setUser(userData.user);
     } catch (err) {
       console.error("Error fetching user:", err);
     }
@@ -49,20 +55,10 @@ const Home = () => {
   // Fetch calendar items from API
   const fetchCalendarItems = async () => {
     try {
-      const response = await fetch(
-        "http://localhost:8080/api/calendarItems/me",
-        {
-          credentials: "include",
-        }
-      );
-
-      if (response.ok) {
-        const items = await response.json();
-        setCalendarItems(items);
-        transformCalendarData(items);
-      } else {
-        setError("Failed to fetch calendar items");
-      }
+      const items = await calendarAPI.getMyItems();
+      console.log("Fetched calendar items:", items); // Debug log
+      setCalendarItems(items);
+      transformAndSetEvents(items);
     } catch (err) {
       setError("Error loading calendar data");
       console.error("Error fetching calendar items:", err);
@@ -71,94 +67,10 @@ const Home = () => {
     }
   };
 
-  // Transform calendar items to TOAST UI format
-  const transformCalendarData = (items) => {
-    console.log("Raw calendar items:", items); // Debug log
-
-    const transformedEvents = items.map((item) => {
-      const calendarId = determineCalendarId(item);
-      console.log(
-        `Item ${item.id}: calendarId = ${calendarId}, event =`,
-        item.event
-      ); // Debug log
-
-      return {
-        id: item.id.toString(),
-        calendarId: calendarId,
-        title: item.title,
-        body: item.description || "",
-        start: new Date(item.start),
-        end: new Date(item.end),
-        location: item.location || "",
-        isAllday: isAllDayEvent(item.start, item.end),
-        category: "time",
-        isVisible: calendarVisibility[calendarId], // Use current visibility setting
-        backgroundColor: getEventColor(item),
-        borderColor: getEventColor(item),
-        raw: item, // Store original data for reference
-      };
-    });
-
+  // Transform and set events
+  const transformAndSetEvents = (items) => {
+    const transformedEvents = transformCalendarData(items, calendarVisibility);
     setEvents(transformedEvents);
-  };
-
-  // Determine calendar category for event
-  const determineCalendarId = (item) => {
-    // Check if this item has an associated event
-    if (item.event) {
-      // If the event is not published, it's a draft
-      if (!item.event.published) {
-        return "drafts";
-      }
-      // If the event has a businessId, it's a business event
-      if (item.event.businessId) {
-        return "business";
-      }
-      // If the event has no businessId, it's a user-created event
-      return "events";
-    }
-    // If no event record, it's a personal calendar item
-    return "personal";
-  };
-
-  // Check if event spans full day
-  const isAllDayEvent = (start, end) => {
-    const startDate = new Date(start);
-    const endDate = new Date(end);
-    const timeDiff = endDate - startDate;
-    const hoursDiff = timeDiff / (1000 * 60 * 60);
-
-    return hoursDiff >= 24;
-  };
-
-  // Determine if event should be visible based on current filters
-  const shouldEventBeVisible = (item) => {
-    const calendarId = determineCalendarId(item);
-
-    // Check privacy for personal items
-    if (calendarId === "personal" && !item.public) {
-      return calendarVisibility.personal;
-    }
-
-    return calendarVisibility[calendarId];
-  };
-
-  // Get event color based on type
-  const getEventColor = (item) => {
-    const calendarId = determineCalendarId(item);
-
-    switch (calendarId) {
-      case "personal":
-        return "#8b5cf6"; // Purple
-      case "business":
-        return "#3b82f6"; // Blue
-      case "events":
-        return "#ec4899"; // Pink
-      case "drafts":
-        return "#6b7280"; // Gray
-      default:
-        return "#6b7280"; // Gray
-    }
   };
 
   // Handle calendar visibility toggle
@@ -207,37 +119,6 @@ const Home = () => {
     }
   };
 
-  // Format current date for display
-  const formatCurrentDate = () => {
-    return currentDate.toLocaleDateString("en-US", {
-      month: "long",
-      year: "numeric",
-    });
-  };
-
-  // Calculate event statistics
-  const getEventStats = () => {
-    const today = new Date();
-    today.setHours(0, 0, 0, 0);
-    const tomorrow = new Date(today);
-    tomorrow.setDate(tomorrow.getDate() + 1);
-
-    const todayEvents = calendarItems.filter((item) => {
-      const eventDate = new Date(item.start);
-      eventDate.setHours(0, 0, 0, 0);
-      return (
-        eventDate.getTime() === today.getTime() && shouldEventBeVisible(item)
-      );
-    });
-
-    const totalVisibleEvents = calendarItems.filter(shouldEventBeVisible);
-
-    return {
-      total: totalVisibleEvents.length,
-      today: todayEvents.length,
-    };
-  };
-
   // Handle event click
   const handleEventClick = (eventInfo) => {
     console.log("Event clicked:", eventInfo);
@@ -258,118 +139,39 @@ const Home = () => {
     setShowCreateModal(true);
   };
 
-  // Create new calendar item
+  // Create new calendar item/event
   const handleCreateEvent = async (eventData) => {
     try {
+      console.log("Creating event with data:", eventData);
+
       // Step 1: Create the calendar item first
-      const response = await fetch(
-        "http://localhost:8080/api/calendarItems/user/item",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          credentials: "include",
-          body: JSON.stringify({
-            title: eventData.title,
-            description: eventData.description,
-            location: eventData.location,
-            start: eventData.start,
-            end: eventData.end,
-            public: eventData.isEvent ? true : eventData.public, // Events are always public
-          }),
-        }
-      );
+      const newCalendarItem = await calendarAPI.createItem({
+        title: eventData.title,
+        description: eventData.description,
+        location: eventData.location,
+        start: eventData.start,
+        end: eventData.end,
+        public: eventData.isEvent ? true : eventData.public, // Events are always public
+      });
 
-      if (response.ok) {
-        const newCalendarItem = await response.json();
+      console.log("Created calendar item:", newCalendarItem);
 
-        // Step 2: If this should be an Event, create the Event record
-        if (eventData.isEvent) {
-          const eventResponse = await fetch(
-            "http://localhost:8080/api/calendarItems/events",
-            {
-              method: "POST",
-              headers: {
-                "Content-Type": "application/json",
-              },
-              credentials: "include",
-              body: JSON.stringify({
-                itemId: newCalendarItem.id,
-                businessId: null,
-                published: eventData.published || false,
-              }),
-            }
-          );
-
-          if (!eventResponse.ok) {
-            console.error("Failed to create event record");
-          }
-        }
-
-        // Refresh calendar data
-        await fetchCalendarItems();
-        setShowCreateModal(false);
-        setSelectedDateTime(null);
-      } else {
-        console.error("Failed to create calendar item");
+      // Step 2: If this should be an Event, create the Event record
+      if (eventData.isEvent) {
+        const eventRecord = await eventsAPI.createEvent({
+          itemId: newCalendarItem.id,
+          businessId: null,
+          published: eventData.published || false,
+        });
+        console.log("Created event record:", eventRecord);
       }
+
+      // Step 3: Refresh calendar data
+      await fetchCalendarItems();
+      setShowCreateModal(false);
+      setSelectedDateTime(null);
     } catch (error) {
       console.error("Error creating event:", error);
-    }
-  };
-
-  // Update calendar item
-  const handleUpdateEvent = async (eventId, eventData) => {
-    try {
-      const response = await fetch(
-        `http://localhost:8080/api/calendarItems/user/item/${eventId}`,
-        {
-          method: "PATCH",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          credentials: "include",
-          body: JSON.stringify(eventData),
-        }
-      );
-
-      if (response.ok) {
-        // Refresh calendar data
-        await fetchCalendarItems();
-        setShowEventModal(false);
-        setSelectedEvent(null);
-      } else {
-        console.error("Failed to update event");
-      }
-    } catch (error) {
-      console.error("Error updating event:", error);
-    }
-  };
-
-  // Delete calendar item
-  const handleDeleteEvent = async (eventId) => {
-    if (window.confirm("Are you sure you want to delete this event?")) {
-      try {
-        const response = await fetch(
-          `http://localhost:8080/api/calendarItems/user/item/${eventId}`,
-          {
-            method: "DELETE",
-            credentials: "include",
-          }
-        );
-
-        if (response.ok) {
-          // Refresh calendar data
-          await fetchCalendarItems();
-          setShowEventModal(false);
-          setSelectedEvent(null);
-        } else {
-          console.error("Failed to delete event");
-        }
-      } catch (error) {
-        console.error("Error deleting event:", error);
-      }
     }
   };
 
@@ -382,52 +184,8 @@ const Home = () => {
     setShowCreateModal(true);
   };
 
-  // Calendar options for TOAST UI
-  const calendarOptions = {
-    defaultView: "month",
-    useFormPopup: false,
-    useDetailPopup: false,
-    calendars: [
-      {
-        id: "personal",
-        name: "Personal",
-        backgroundColor: "#8b5cf6",
-        borderColor: "#8b5cf6",
-        dragBackgroundColor: "#8b5cf6",
-      },
-      {
-        id: "business",
-        name: "Business",
-        backgroundColor: "#3b82f6",
-        borderColor: "#3b82f6",
-        dragBackgroundColor: "#3b82f6",
-      },
-      {
-        id: "events",
-        name: "Events",
-        backgroundColor: "#ec4899",
-        borderColor: "#ec4899",
-        dragBackgroundColor: "#ec4899",
-      },
-      {
-        id: "drafts",
-        name: "Drafts",
-        backgroundColor: "#6b7280",
-        borderColor: "#6b7280",
-        dragBackgroundColor: "#6b7280",
-      },
-    ],
-    week: {
-      startDayOfWeek: 0,
-      dayNames: ["S", "M", "T", "W", "T", "F", "S"],
-      hourStart: 0,
-      hourEnd: 24,
-    },
-    month: {
-      startDayOfWeek: 0,
-      dayNames: ["S", "M", "T", "W", "T", "F", "S"],
-    },
-  };
+  // Get calendar options
+  const calendarOptions = getCalendarOptions();
 
   // Initialize data on component mount
   useEffect(() => {
@@ -435,14 +193,15 @@ const Home = () => {
     fetchCalendarItems();
   }, []);
 
+  // Force month view on mount
   useEffect(() => {
     if (calendarRef.current) {
       const calendar = calendarRef.current.getInstance();
-      calendar.changeView("month"); // Force month view on mount
+      calendar.changeView("month");
     }
-  }, []); // Run once when component mounts
+  }, []);
 
-  // Update calendar events when data changes
+  // Update calendar events when data or visibility changes
   useEffect(() => {
     if (calendarRef.current && events.length > 0) {
       const calendar = calendarRef.current.getInstance();
@@ -459,7 +218,14 @@ const Home = () => {
     }
   }, [events, calendarVisibility]);
 
-  const stats = getEventStats();
+  // Update events when visibility changes
+  useEffect(() => {
+    if (calendarItems.length > 0) {
+      transformAndSetEvents(calendarItems);
+    }
+  }, [calendarVisibility]);
+
+  const stats = getEventStats(calendarItems, calendarVisibility);
 
   if (loading) {
     return (
@@ -484,7 +250,7 @@ const Home = () => {
         {/* Create Events Button */}
         <button className="create-events-btn" onClick={handleCreateEventsClick}>
           <span className="plus-icon">+</span>
-          Create Events
+          Create Event
         </button>
 
         {/* My Network Section */}
@@ -611,7 +377,7 @@ const Home = () => {
             >
               ‹
             </button>
-            <h2 className="current-date">{formatCurrentDate()}</h2>
+            <h2 className="current-date">{formatCurrentDate(currentDate)}</h2>
             <button
               className="nav-btn"
               onClick={() => handleNavigation("next")}
@@ -671,9 +437,7 @@ const Home = () => {
             setShowEventModal(false);
             setSelectedEvent(null);
           }}
-          onUpdate={handleUpdateEvent}
-          onDelete={handleDeleteEvent}
-          fetchCalendarItems={fetchCalendarItems}
+          onRefresh={fetchCalendarItems}
         />
       )}
 
@@ -688,435 +452,6 @@ const Home = () => {
           onCreate={handleCreateEvent}
         />
       )}
-    </div>
-  );
-};
-
-// Event Detail Modal Component
-const EventDetailModal = ({
-  event,
-  onClose,
-  onUpdate,
-  onDelete,
-  fetchCalendarItems,
-}) => {
-  const [isEditing, setIsEditing] = useState(false);
-  const [formData, setFormData] = useState({
-    title: event.title || "",
-    description: event.description || "",
-    location: event.location || "",
-    start: new Date(event.start).toISOString().slice(0, 16),
-    end: new Date(event.end).toISOString().slice(0, 16),
-    public: event.public || false,
-  });
-
-  // Determine if this is an event (has event record)
-  const isEvent = event.event !== undefined;
-  const isPublished = isEvent && event.event?.published;
-
-  const handleInputChange = (e) => {
-    const { name, value, type, checked } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: type === "checkbox" ? checked : value,
-    }));
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    onUpdate(event.id, {
-      ...formData,
-      start: new Date(formData.start).toISOString(),
-      end: new Date(formData.end).toISOString(),
-    });
-  };
-
-  const handlePublishToggle = async () => {
-    try {
-      const response = await fetch(
-        `http://localhost:8080/api/calendarItems/events/${event.event.id}`,
-        {
-          method: "PATCH",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          credentials: "include",
-          body: JSON.stringify({
-            published: !isPublished,
-          }),
-        }
-      );
-
-      if (response.ok) {
-        // Refresh calendar data instead of full page reload
-        onClose(); // Close modal first
-        await fetchCalendarItems(); // Refresh data
-      } else {
-        console.error("Failed to toggle publish status");
-      }
-    } catch (error) {
-      console.error("Error toggling publish status:", error);
-    }
-  };
-
-  return (
-    <div className="modal-overlay" onClick={onClose}>
-      <div className="modal-content" onClick={(e) => e.stopPropagation()}>
-        <div className="modal-header">
-          <div>
-            <h2>{isEditing ? "Edit Event" : "Event Details"}</h2>
-            {isEvent && (
-              <span
-                className={`status-badge ${
-                  isPublished ? "published" : "draft"
-                }`}
-              >
-                {isPublished ? "🌐 Published" : "📝 Draft"}
-              </span>
-            )}
-          </div>
-          <button className="close-btn" onClick={onClose}>
-            ×
-          </button>
-        </div>
-
-        {isEditing ? (
-          <form onSubmit={handleSubmit} className="event-form">
-            <div className="form-group">
-              <label>Title *</label>
-              <input
-                type="text"
-                name="title"
-                value={formData.title}
-                onChange={handleInputChange}
-                required
-              />
-            </div>
-            <div className="form-group">
-              <label>Description</label>
-              <textarea
-                name="description"
-                value={formData.description}
-                onChange={handleInputChange}
-                rows={3}
-              />
-            </div>
-            <div className="form-group">
-              <label>Location</label>
-              <input
-                type="text"
-                name="location"
-                value={formData.location}
-                onChange={handleInputChange}
-              />
-            </div>
-            <div className="form-row">
-              <div className="form-group">
-                <label>Start Date & Time *</label>
-                <input
-                  type="datetime-local"
-                  name="start"
-                  value={formData.start}
-                  onChange={handleInputChange}
-                  required
-                />
-              </div>
-
-              <div className="form-group">
-                <label>End Date & Time *</label>
-                <input
-                  type="datetime-local"
-                  name="end"
-                  value={formData.end}
-                  onChange={handleInputChange}
-                  required
-                />
-              </div>
-            </div>
-            {!isEvent && (
-              <div className="form-group">
-                <label className="checkbox-label">
-                  <input
-                    type="checkbox"
-                    name="public"
-                    checked={formData.public}
-                    onChange={handleInputChange}
-                  />
-                  Make this visible to friends
-                </label>
-              </div>
-            )}
-            <div className="modal-actions">
-              <button
-                type="button"
-                onClick={() => setIsEditing(false)}
-                className="btn-secondary"
-              >
-                Cancel
-              </button>
-              <button type="submit" className="btn-primary">
-                Save Changes
-              </button>
-            </div>
-          </form>
-        ) : (
-          <div className="event-details">
-            <div className="detail-item">
-              <strong>Title:</strong>
-              <span>{event.title}</span>
-            </div>
-
-            {event.description && (
-              <div className="detail-item">
-                <strong>Description:</strong>
-                <span>{event.description}</span>
-              </div>
-            )}
-
-            {event.location && (
-              <div className="detail-item">
-                <strong>Location:</strong>
-                <span>{event.location}</span>
-              </div>
-            )}
-
-            <div className="detail-item">
-              <strong>Start:</strong>
-              <span>{new Date(event.start).toLocaleString()}</span>
-            </div>
-
-            <div className="detail-item">
-              <strong>End:</strong>
-              <span>{new Date(event.end).toLocaleString()}</span>
-            </div>
-
-            <div className="detail-item">
-              <strong>Type:</strong>
-              <span>{isEvent ? "Public Event" : "Personal Calendar Item"}</span>
-            </div>
-
-            {!isEvent && (
-              <div className="detail-item">
-                <strong>Visibility:</strong>
-                <span>{event.public ? "Visible to friends" : "Private"}</span>
-              </div>
-            )}
-
-            <div className="modal-actions">
-              <button onClick={() => onDelete(event.id)} className="btn-danger">
-                Delete {isEvent ? "Event" : "Item"}
-              </button>
-              {isEvent && (
-                <button onClick={handlePublishToggle} className="btn-secondary">
-                  {isPublished ? "Unpublish" : "Publish"} Event
-                </button>
-              )}
-              <button
-                onClick={() => setIsEditing(true)}
-                className="btn-primary"
-              >
-                Edit {isEvent ? "Event" : "Item"}
-              </button>
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
-
-// Create Event Modal Component
-const CreateEventModal = ({ selectedDateTime, onClose, onCreate }) => {
-  const [formData, setFormData] = useState({
-    title: "",
-    description: "",
-    location: "",
-    start: selectedDateTime
-      ? new Date(selectedDateTime.start).toISOString().slice(0, 16)
-      : "",
-    end: selectedDateTime
-      ? new Date(selectedDateTime.end).toISOString().slice(0, 16)
-      : "",
-    public: false,
-    isEvent: false, // New field to determine if this should be an Event
-  });
-
-  const handleInputChange = (e) => {
-    const { name, value, type, checked } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: type === "checkbox" ? checked : value,
-    }));
-  };
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    onCreate({
-      ...formData,
-      start: new Date(formData.start).toISOString(),
-      end: new Date(formData.end).toISOString(),
-      published: true, // Regular submit publishes event
-    });
-  };
-
-  return (
-    <div className="modal-overlay" onClick={onClose}>
-      <div className="modal-content" onClick={(e) => e.stopPropagation()}>
-        <div className="modal-header">
-          <h2>Create New {formData.isEvent ? "Event" : "Calendar Item"}</h2>
-          <button className="close-btn" onClick={onClose}>
-            ×
-          </button>
-        </div>
-
-        <form onSubmit={handleSubmit} className="event-form">
-          {/* Event Type Selection */}
-          <div className="form-group">
-            <label className="form-section-title">What are you creating?</label>
-            <div className="radio-group">
-              <label className="radio-label">
-                <input
-                  type="radio"
-                  name="isEvent"
-                  checked={!formData.isEvent}
-                  onChange={() =>
-                    setFormData((prev) => ({ ...prev, isEvent: false }))
-                  }
-                />
-                <span>Personal Calendar Item</span>
-                <small>Private reminder or personal task</small>
-              </label>
-              <label className="radio-label">
-                <input
-                  type="radio"
-                  name="isEvent"
-                  checked={formData.isEvent}
-                  onChange={() =>
-                    setFormData((prev) => ({ ...prev, isEvent: true }))
-                  }
-                />
-                <span>Public Event</span>
-                <small>Others can discover and attend this event</small>
-              </label>
-            </div>
-          </div>
-
-          <div className="form-group">
-            <label>Title *</label>
-            <input
-              type="text"
-              name="title"
-              value={formData.title}
-              onChange={handleInputChange}
-              required
-              placeholder={
-                formData.isEvent ? "Event title" : "Calendar item title"
-              }
-            />
-          </div>
-
-          <div className="form-group">
-            <label>Description {formData.isEvent ? "*" : ""}</label>
-            <textarea
-              name="description"
-              value={formData.description}
-              onChange={handleInputChange}
-              required={formData.isEvent}
-              rows={3}
-              placeholder={
-                formData.isEvent
-                  ? "Event description (required for events)"
-                  : "Optional description"
-              }
-            />
-          </div>
-
-          <div className="form-group">
-            <label>Location {formData.isEvent ? "*" : ""}</label>
-            <input
-              type="text"
-              name="location"
-              value={formData.location}
-              onChange={handleInputChange}
-              required={formData.isEvent}
-              placeholder={
-                formData.isEvent
-                  ? "Event location (required for events)"
-                  : "Optional location"
-              }
-            />
-          </div>
-
-          <div className="form-row">
-            <div className="form-group">
-              <label>Start Date & Time *</label>
-              <input
-                type="datetime-local"
-                name="start"
-                value={formData.start}
-                onChange={handleInputChange}
-                required
-              />
-            </div>
-
-            <div className="form-group">
-              <label>End Date & Time *</label>
-              <input
-                type="datetime-local"
-                name="end"
-                value={formData.end}
-                onChange={handleInputChange}
-                required
-              />
-            </div>
-          </div>
-
-          {!formData.isEvent && (
-            <div className="form-group">
-              <label className="checkbox-label">
-                <input
-                  type="checkbox"
-                  name="public"
-                  checked={formData.public}
-                  onChange={handleInputChange}
-                />
-                Make this visible to friends
-              </label>
-            </div>
-          )}
-
-          <div className="modal-actions">
-            <button type="button" onClick={onClose} className="btn-secondary">
-              Cancel
-            </button>
-            {formData.isEvent ? (
-              <>
-                <button
-                  type="button"
-                  className="btn-secondary"
-                  onClick={() => {
-                    onCreate({
-                      ...formData,
-                      start: new Date(formData.start).toISOString(),
-                      end: new Date(formData.end).toISOString(),
-                      published: false, // Save as draft
-                    });
-                  }}
-                >
-                  Save as Draft
-                </button>
-                <button type="submit" className="btn-primary">
-                  Publish Event
-                </button>
-              </>
-            ) : (
-              <button type="submit" className="btn-primary">
-                Create Calendar Item
-              </button>
-            )}
-          </div>
-        </form>
-      </div>
     </div>
   );
 };
